refactor: group props in video controls; usePlayerKeyboard
